코드 정적 분석 리팩토링
무분별한 세미콜론 사용과 일관성을 해치는 tab/space blank 사용, 비직관적/비효율적 if condition 개선, 배열의 자료형 선언 시 []가 자료형 바로 뒤에 오도록 일관성 추구

+++ core/src/main/java/com/gfpixel/gfpixeldungeon/levels/features/HighGrass.java
@@ -63,7 +63,7 @@ public class HighGrass {
 
 		if (naturalismLevel >= 0) {
 			// Seed, scales from 1/16 to 1/4
-			if (Random.Int(16 - ((int) (naturalismLevel * 3))) == 0) {
+			if (Random.Int(16 - naturalismLevel * 3) == 0) {
 				Item seed = Generator.random(Generator.Category.SEED);
 
 				if (seed instanceof BlandfruitBush.Seed) {
